requirements.txt
python-telegram-bot==13.12
environs==9.5.0
redis==3.2.1

